# swagger.yaml
openapi: 3.0.0
info:
  title: jobs-api
  contact: {}
  version: "1.0"
servers:
  - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
    variables: {}
  - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
    variables: {}
paths:
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register
      operationId: Register
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              name: okay123
              email: okay@gmail.com
              password: "1234567"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
      servers:
        - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
          variables: {}
  /auth/login:
    post:
      tags:
        - Auth
      summary: login
      operationId: login
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/loginrequest"
            example:
              email: david@gmail.com
              password: "654321"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
      servers:
        - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
          variables: {}
  /jobs:
    post:
      tags:
        - Jobs
      summary: Create Job
      operationId: CreateJob
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
            example:
              company: apple
              position: random
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
          variables: {}
    get:
      tags:
        - Jobs
      summary: Get all Jobs
      operationId: GetallJobs
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
          variables: {}
  /jobs/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: The user ID
    get:
      tags:
        - Jobs
      summary: Get a Job
      operationId: GetaJob
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
          variables: {}
    patch:
      tags:
        - Jobs
      summary: Update Job
      operationId: UpdateJob
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
            example:
              company: random
              position: nothing
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
          variables: {}
    delete:
      tags:
        - Jobs
      summary: Delete Job
      operationId: DeleteJob
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security:
        - bearer: []
      servers:
        - url: https://jobs-api-08-06-02.herokuapp.com/api/v1
          variables: {}
components:
  schemas:
    RegisterRequest:
      title: RegisterRequest
      required:
        - name
        - email
        - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: okay123
        email: okay@gmail.com
        password: "1234567"
    loginrequest:
      title: loginrequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: david@gmail.com
        password: "654321"
    CreateJobRequest:
      title: CreateJobRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: apple
        position: random
    UpdateJobRequest:
      title: UpdateJobRequest
      required:
        - company
        - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: random
        position: nothing
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
tags:
  - name: Misc
    description: ""
  - name: Auth
    description: ""
  - name: Jobs
    description: ""